refactor: Enums relocate and/or rename to singular/plural (#386)
* EPoints -> EPoint

* EGuildJoinStatusCode file rename

* EPhases -> EPhase

* ERaceFlag -> ERaceFlags

* ESkillFlag -> ESkillFlags

* ESkillIndexes.cs => ESkill.cs

* ESkillAffectFlag.cs => EAffectFlags.cs

* EPremiumTypes.cs => EPremiumType.cs

* ChatMessageTypes.cs => ChatMessageType.cs

* EquipmentSlots.cs => EquipmentSlot.cs

* ISKill.cs => ISkill.cs

* EMapAttribute.cs => EMapAttributes.cs

* new Types/Combat namespace

* new Types/Monsters namespace

* new Types/Items namespace

* new Types/Players namespace

* new Types/Entities namespace

* new Types/Guild namespace

* new Types/Quest namespace

* IPlayer file rename to match

* use only primitive types in packet definitions

* Revert "use only primitive types in packet definitions"

This reverts commit 3df318d66b0fbdb5b1524549b107baa83131f50e.

* Use enum types for (most) closed set packet fields

* [JetBrains Rider] Remove unused directives in solution

+++ src/Data/Game.Persistence/DbPlayerRepository.cs
@@ -41,7 +41,7 @@ public class DbPlayerRepository : IDbPlayerRepository
             AccountId = player.AccountId,
             Name = player.Name,
             PlayerClass = (byte)player.PlayerClass,
-            SkillGroup = player.SkillGroup,
+            SkillGroup = (byte)player.SkillGroup,
             PlayTime = player.PlayTime,
             Level = player.Level,
             Experience = player.Experience,
@@ -88,7 +88,7 @@ public class DbPlayerRepository : IDbPlayerRepository
 
         entity.Empire = data.Empire;
         entity.PlayerClass = (byte)data.PlayerClass;
-        entity.SkillGroup = data.SkillGroup;
+        entity.SkillGroup = (byte)data.SkillGroup;
         entity.PlayTime = data.PlayTime;
         entity.Level = data.Level;
         entity.Experience = data.Experience;
